refactor: remove unnecessary indirect event processing
I thought that this indirection (storing the events after the browser receives interaction and triggers an event handler, in a queue (implemented using an array), and processing them all at once at the start of the next call to `requestAnimationFrame`) was necessary as I thought that event handlers could run at any time, even while a `requestAnimationFrame` callback (rendering code in this case). If this was true, it would mean that values could be updated mid-render and thus lead to unpredictable results. But after looking this up (along with getting more info on Javascript's execution model), it seems that event handlers will be called before the next scheduled `requestAnimationFrame` callback, therefore the approach I had before was not needed in the first place. This removal was also done (but not acknowledged) while adding the `movable-camera` scene.