refactor: remove `shaderSources` parameter from `RenderWrapper`
Removing the `shaderSources` parameter invalidates the need for the `getShaderSources` function which was a convenience over manually importing the shader files with "?raw" at the top of each scene file. `getShaderSources` assumed that each scene only creates one shader program (i.e. one vertex and fragment shader pair for one scene object type), and reverting back to the manual import allows the scene to determine the number of shaders it needs.

The implications of having `RenderWrapper` only take in the `canvas` has been listed in the inserted JSDoc comments for the type.